+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog of PicoScenes MATLAB Toolbox (PMT)
 
-## Jan. 26, 2021
+## Feb. 16, 2022
+    * [Parser] Add parser for PicoScenes .bbsignals file (baseband signal recording format used by PicoScenes).
+        .bbsignals files can be decoded by double-click in MATLAB Current Folder or drag'n'drop.
+
+## Jan. 26, 2022
     * [Rxs-Parsing-Core] CSD removal is now integrated in parsing-core.
 
 ## Jan. 18, 2022